refactor(agents): 抽离版本解析辅助函数到 utils 模块并补充 AssistantStoreTab 类型注释
